# openapi-live.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "s-curve",
    "description": "",
    "contact": {
      "name": "s-curve team"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8800"
    },
    {
      "url": "https://rust-service:8800"
    }
  ],
  "paths": {
...
    "/telemetry/events": {
      "post": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Ingest Events",
        "description": "Ingest Events Handles `POST` requests for `/telemetry/events`. Requires bearer authentication.",
        "operationId": "ingest_events",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted telemetry events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryIngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid telemetry payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Ingest Events Handles `POST` requests for `/telemetry/events`. Requires bearer authentication.",
        "operationId": "ingest_events",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryIngestResponse"
                }
              }
            },
            "description": "Accepted telemetry events"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryErrorResponse"
                }
              }
            },
            "description": "Invalid telemetry payload"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Ingest Events",
        "tags": [
          "Telemetry"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TelemetryBatchRequest": {
        "type": "object",
        "required": [
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TelemetryEventRequest"
            },
            "maxItems": 20,
            "minItems": 1
          }
        }
      },
      "TelemetryErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "invalid telemetry payload"
          }
        }
      },
      "TelemetryEventName": {
        "type": "string",
        "enum": [
          "time_to_task.session_started",
          "time_to_task.intent_marked",
          "time_to_task.completed",
          "time_to_task.abandoned"
        ]
      },
      "TelemetryEventRequest": {
        "type": "object",
        "required": [
          "event_id",
          "event_name",
          "occurred_at",
          "metadata"
        ],
        "properties": {
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "example": 9234,
            "nullable": true,
            "minimum": 0
          },
          "event_id": {
            "type": "string",
            "example": "time-to-task-start-c6f53fe6-1823-4f90-b03f-77e8e20de34d"
          },
          "event_name": {
            "$ref": "#/components/schemas/TelemetryEventName"
          },
          "intent_to_complete_ms": {
            "type": "integer",
            "format": "int64",
            "example": 4033,
            "nullable": true,
            "minimum": 0
          },
          "metadata": {
            "type": "object"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-04T05:38:12.232Z"
          },
          "outcome": {
            "type": "string",
            "example": "completed",
            "nullable": true
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "example": "leave-tasks-page",
            "nullable": true
          },
          "route": {
            "type": "string",
            "example": "/tasks",
            "nullable": true
          },
          "session_id": {
            "type": "string",
            "example": "c6f53fe6-1823-4f90-b03f-77e8e20de34d",
            "nullable": true
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "view": {
            "type": "string",
            "example": "list",
            "nullable": true
          }
        }
      },
      "TelemetryIngestResponse": {
        "type": "object",
        "required": [
          "accepted"
        ],
        "properties": {
          "accepted": {
            "type": "integer",
            "example": 12,
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}